social_app/social_app_1/requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.2.0
PyMuPDF==1.23.8
pandas==2.2.0
streamlit==1.31.1
SQLAlchemy==2.0.27
python-dotenv==1.0.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
chromadb==0.4.22
ollama==0.1.6
psycopg2-binary==2.9.9 